perf: seperate logic to ui in chat box and create custom hook
